fetal-brain-measurement/Code/FetalMeasurements-master/requirements.txt
absl-py==0.9.0
ansi2html==1.5.2
astor==0.8.1
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
blis==0.4.1
Bottleneck==1.3.1
brotlipy==0.7.0
cachetools==4.2.2
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
CHAID==5.3.0
chardet==3.0.4
circle-fit==0.1.3
Click==7.0
colorama==0.4.3
colorlover==0.3.0
configparser==5.0.0
contextvars==2.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
cytoolz==0.11.0
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
dicom2nifti==2.2.10
dill==0.3.4
diskcache==5.4.0
docopt==0.6.2
easydict==1.7
entrypoints==0.3
et-xmlfile==1.0.1
etelemetry==0.1.2
fastai==1.0.60
fastcore==1.4.0
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.0.12
future==0.18.2
future-fstrings==1.0.0
gast==0.3.3
gitdb2==3.0.2
GitPython==3.0.8
google-pasta==0.1.8
graphviz==0.16
grpcio==1.26.0
h5py==2.10.0
hdf5storage==0.1.15
idna==2.9
imagecodecs==2020.2.18
imageio==2.6.1
imbalanced-learn==0.7.0
imgaug==0.4.0
immutables==0.19
importlib-metadata==4.8.3
incense==0.0.10
iniconfig==1.1.1
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json-tricks==3.15.2
json5==0.9.1
jsonpickle==1.3
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
littleutils==0.2.2
llvmlite==0.35.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.4
MedPy==0.4.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mlxtend==0.18.0
mpmath==1.1.0
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.1
NiftyNet==0.2.0
nilearn==0.6.1
ninja==1.10.0.post1
nipype==1.4.2
notebook==6.0.3
numba==0.52.0
numexpr==2.7.1
numpy==1.19.4
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==3.4.1.15
openpyxl==3.0.3
outdated==0.2.1
packaging==20.1
pandas==1.0.1
pandas-flavor==0.2.0
pandocfilters==1.4.2
parso==0.6.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pingouin==0.3.11
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.2
prov==1.5.3
psutil==5.9.4
ptyprocess==0.6.0
py==1.10.0
py-cpuinfo==5.0.0
pydicom==1.4.2
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.5.2
pykwalify==1.8.0
pyparsing==2.4.6
pyprof2calltree==1.4.5
pyradiomics==3.0.1
pyrsistent==0.15.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
rdflib==4.2.2
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sacred==0.8.1
savReaderWriter==3.4.2
scikit-fuzzy==0.4.2
scikit-image==0.16.2
scikit-learn==0.23.0
scipy==1.2.2
seaborn==0.11.1
Send2Trash==1.5.0
shap==0.35.0
Shapely==1.6.4
SimpleITK==1.2.4
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smmap2==2.0.5
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
statsmodels==0.11.0
sympy==1.6.2
tables==3.6.1
tabulate==0.8.9
tensorboard==1.14.0
tensorboardX==2.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
threadpoolctl==2.1.0
tifffile==2020.2.16
tokenize-rt==2.2.0
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
traits==6.0.0
treelib==1.6.1
typing-extensions==4.1.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.0
xarray==0.16.2
xgboost==0.80
xlrd==1.2.0
yacs==0.1.6
zipp==3.0.0
